+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.1.0 Under development
 
 - Enh #1: Initial commit (@terabytesoftw)
+- Enh #2: Add dropdown cookbook selector `language` and toggle cookbok `selectorLanguage` classes (@terabytesoftw)